Rename `user` related acceptance test directory to `user managment`
